+++ src/pages/_routes/PublicRoutes.jsx
@@ -55,7 +55,28 @@ export const publicRoutesPaths = [
         path: "/productos/:page",
         element: <Products />,
       },
-      { path: "/producto/:id", element: <DetailProduct /> },
+      {
+        path: "/producto/:id",
+        loader: async ({ params }) => {
+          try {
+            const data = (await APISpot.product.getOne({ id: params.id })).data;
+            const map = {};
+            let variants = [];
+            data.variants.map((variant, description) => {
+              const { subRub } = variant;
+              if (!map[subRub]) {
+                map[subRub] = description;
+                variants.push(variant);
+              }
+            });
+            data.variants = variants;
+            return data;
+          } catch (e) {
+            return null;
+          }
+        },
+        element: <DetailProduct />,
+      },
     ],
   },
   {
